requirements.txt
pillow>=10.0
rembg>=2.0
numpy>=1.26
scikit-image>=0.22
scipy>=1.11
onnxruntime